+++ stdlib/kokorawr/kokorawr/functions/services/bs.js
