+++ stdlib/kokorawr/kokorawr/functions/services/ddg.js
