+++ stdlib/kokorawr/kokorawr/functions/services/mm.js
