+++ stdlib/kokorawr/kokorawr/functions/services/rps.js
